Refactors Ethereum clients into separate React component
